requirements.txt
celery
numpy
bokeh
redis
scipy
matplotlib
pillow
